feat: Add "Create Blog" link to Navbar
This commit adds a new link to the Navbar component, allowing users to navigate to the "Create Blog" page. The link is associated with the "/create-blog" URL and has the label "Create Blog". This enhancement improves the navigation experience of the application by providing a direct way to access the page for creating new blog posts.