feat: Update package dependencies
This commit updates the package.json file to include the latest versions of the "dompurify" and "firebase" dependencies. The "dompurify" dependency is updated to version 3.1.6, and the "firebase" dependency is updated to version 10.12.3. These updates ensure that the project is using the latest stable versions of these dependencies, which may include bug fixes, performance improvements, and new features.